+++ keycloak/openid_hardcoded_claim_protocol_mapper.go
@@ -107,11 +107,7 @@ func (keycloakClient *KeycloakClient) UpdateOpenIdHardcodedClaimProtocolMapper(c
 }
 
 func (keycloakClient *KeycloakClient) ValidateOpenIdHardcodedClaimProtocolMapper(ctx context.Context, mapper *OpenIdHardcodedClaimProtocolMapper) error {
-	if mapper.ClientId == "" && mapper.ClientScopeId == "" {
-		return fmt.Errorf("validation error: one of ClientId or ClientScopeId must be set")
-	}
-
-	protocolMappers, err := keycloakClient.listGenericProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
+	protocolMappers, err := keycloakClient.listProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
 	if err != nil {
 		return err
 	}