+++ keycloak/openid_user_realm_role_protocol_mapper.go
@@ -123,11 +123,7 @@ func (keycloakClient *KeycloakClient) UpdateOpenIdUserRealmRoleProtocolMapper(ct
 }
 
 func (keycloakClient *KeycloakClient) ValidateOpenIdUserRealmRoleProtocolMapper(ctx context.Context, mapper *OpenIdUserRealmRoleProtocolMapper) error {
-	if mapper.ClientId == "" && mapper.ClientScopeId == "" {
-		return fmt.Errorf("validation error: one of ClientId or ClientScopeId must be set")
-	}
-
-	protocolMappers, err := keycloakClient.listGenericProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
+	protocolMappers, err := keycloakClient.listProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
 	if err != nil {
 		return err
 	}