+++ keycloak/saml_user_attribute_protocol_mapper.go
@@ -92,11 +92,7 @@ func (keycloakClient *KeycloakClient) UpdateSamlUserAttributeProtocolMapper(ctx 
 }
 
 func (keycloakClient *KeycloakClient) ValidateSamlUserAttributeProtocolMapper(ctx context.Context, mapper *SamlUserAttributeProtocolMapper) error {
-	if mapper.ClientId == "" && mapper.ClientScopeId == "" {
-		return fmt.Errorf("validation error: one of ClientId or ClientScopeId must be set")
-	}
-
-	protocolMappers, err := keycloakClient.listGenericProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
+	protocolMappers, err := keycloakClient.listProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
 	if err != nil {
 		return err
 	}