+++ keycloak/saml_user_property_protocol_mapper.go
@@ -83,11 +83,7 @@ func (keycloakClient *KeycloakClient) UpdateSamlUserPropertyProtocolMapper(ctx c
 }
 
 func (keycloakClient *KeycloakClient) ValidateSamlUserPropertyProtocolMapper(ctx context.Context, mapper *SamlUserPropertyProtocolMapper) error {
-	if mapper.ClientId == "" && mapper.ClientScopeId == "" {
-		return fmt.Errorf("validation error: one of ClientId or ClientScopeId must be set")
-	}
-
-	protocolMappers, err := keycloakClient.listGenericProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
+	protocolMappers, err := keycloakClient.listProtocolMappers(ctx, mapper.RealmId, mapper.ClientId, mapper.ClientScopeId)
 	if err != nil {
 		return err
 	}