+++ provider/resource_keycloak_generic_protocol_mapper_test.go
@@ -76,6 +76,12 @@ func TestAccKeycloakGenericProtocolMapper_import(t *testing.T) {
 				ImportStateVerify: true,
 				ImportStateIdFunc: getGenericProtocolMapperIdForClient(resourceName),
 			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateIdFunc: getGenericProtocolMapperNameImportIdForClient(resourceName),
+			},
 		},
 	})
 }