+++ provider/resource_keycloak_saml_user_property_protocol_mapper_test.go
@@ -56,6 +56,12 @@ func TestAccKeycloakSamlUserPropertyProtocolMapper_import(t *testing.T) {
 				ImportStateVerify: true,
 				ImportStateIdFunc: getGenericProtocolMapperIdForClient(clientResourceName),
 			},
+			{
+				ResourceName:      clientResourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateIdFunc: getGenericProtocolMapperNameImportIdForClient(clientResourceName),
+			},
 		},
 	})
 }